Refactor network code
